requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.2.1
    # via passlib
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   uvicorn
cryptography==44.0.0
    # via
    #   -r requirements.in
    #   pyjwt
fastapi==0.117.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   respx
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
packaging==25.0
    # via pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyjwt[crypto]==2.8.0
    # via -r requirements.in
pytest==8.4.2
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
respx==0.22.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via fastapi
tenacity==8.2.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.37.0
    # via -r requirements.in
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
